added clean bin

+++ source/deploy.bat
@@ -1,6 +1,6 @@
 @echo off
 
-rmdir /s /q bin
+if exist "bin" rmdir /s /q bin
 
 :: Get date with this template =>mouth day hours minutes seconds
 set DATE=%date:~-10,2%%date:~-7,2%%time:~-11,2%%time:~-8,2%%time:~-5,2%
